requirements.txt
ultralytics==8.0.168       
opencv-python              
flask                      
numpy                      
pillow                     
requests                   
matplotlib                 
torch                      
torchvision               

